veo3_mcp/requirements.txt
# Core dependencies
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Google Cloud dependencies
google-cloud-aiplatform>=1.40.0
google-generativeai>=0.3.0
google-cloud-storage>=2.10.0

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0

# Async and networking
httpx>=0.24.0
aiofiles>=23.0.0
websockets>=11.0.0

# OpenTelemetry
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-instrumentation-httpx>=0.41b0
opentelemetry-instrumentation-sqlalchemy>=0.41b0
opentelemetry-instrumentation-logging>=0.41b0
opentelemetry-exporter-otlp-proto-grpc>=1.20.0

# Development dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
mypy>=1.5.0
ruff>=0.0.290